requirements.txt
Flask==2.1.0
gunicorn==20.1.0
tabula-py==2.3.0
google-cloud-vision==2.7.2
google-cloud-storage==2.3.0
google-cloud-bigquery==3.0.1
